@@ -1,50 +1,43 @@
 {
-  "name": "angular-cli",
+  "name": "angularcli",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config proxy.config.json --disable-host-check --open",
+    "local": "ng serve --proxy-config proxy.config.json --ssl --ssl-key ca/localhost.key --ssl-cert ca/localhost.crt",
+    "dev": "ng serve --proxy-config proxy.config.json --disable-host-check",
+    "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.1.4",
-    "@angular/common": "12.1.4",
-    "@angular/compiler": "12.1.4",
-    "@angular/core": "12.1.4",
-    "@angular/forms": "12.1.4",
-    "@angular/http": "7.2.16",
-    "@angular/platform-browser": "12.1.4",
-    "@angular/platform-browser-dynamic": "12.1.4",
-    "@angular/router": "12.1.4",
-    "core-js": "3.8.3",
-    "rxjs": "6.6.3",
+    "@angular/animations": "~12.2.0",
+    "@angular/common": "~12.2.0",
+    "@angular/compiler": "~12.2.0",
+    "@angular/core": "~12.2.0",
+    "@angular/forms": "~12.2.0",
+    "@angular/platform-browser": "~12.2.0",
+    "@angular/platform-browser-dynamic": "~12.2.0",
+    "@angular/router": "~12.2.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.3.0",
+    "uuid": "^8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "12.1.4",
-    "@angular/cli": "12.1.4",
-    "@angular/compiler-cli": "12.1.4",
-    "@angular/language-service": "12.1.4",
-    "@types/jasmine": "3.6.3",
-    "@types/jasminewd2": "2.0.8",
-    "@types/node": "14.14.28",
-    "codelyzer": "6.0.1",
-    "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "6.0.0",
-    "karma": "6.3.4",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.5.4",
-    "protractor": "7.0.0",
-    "ts-node": "9.1.1",
-    "tslint": "6.1.3",
-    "typescript": "4.3.5"
-  },
-  "keywords": ["angular"],
-  "description": "An Angular starter you can use ng commands with to create components, routes, services, pipes etc."
+    "@angular/language-service": "~12.2.0",
+    "@angular-devkit/build-angular": "~12.2.0",
+    "@angular/cli": "~12.2.0",
+    "@angular/compiler-cli": "~12.2.0",
+    "@types/jasmine": "~3.8.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~3.8.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.3.5"
+  }
 }
