@@ -1,17 +1,21 @@
 import { Assert } from "../helper/Assert";
-import { LeafNode } from "./LeafNode";
+import { LeafNode } from "./1006_LeafNode";
 import { StringBuilder } from "../helper/StringBuilder";
 import { OutputSetting } from "../parse/Setting";
-import { NodeUtils } from "../helper/NodeUtils";
 import { Element } from "./Element";
 import { StringUtil } from "../helper/StringUtil";
 import { Objects } from "../helper/Objects";
 import { Entities } from "./Entities";
+import { Node } from "./1004_Node";
+import { Document } from "./Document";
 
 /**
  * A text node.
  */
 export class TextNode extends LeafNode {
+  static is(node: Node): node is TextNode {
+    return node instanceof TextNode;
+  }
   /**
    * Create a new TextNode representing the supplied (unencoded) text).
    * @param text raw text
@@ -109,13 +113,13 @@ export class TextNode extends LeafNode {
     let siblingIndex = this.getSiblingIndex();
 
     // if pretty === true
-    let isIf0 = () => siblingIndex === 0 &&  NodeUtils.isElement(parent) &&  parent.tag().formatAsBlock && !this.isBlank;
+    let isIf0 = () => siblingIndex === 0 &&  Element.is(parent) &&  parent.tag().formatAsBlock && !this.isBlank;
     let isIf1 = () => setting.outline && this.siblingNodes().length > 0 && !this.isBlank();
     if (pretty && isIf0() && isIf1()) this.indent(accum, depth, setting);
 
     //
     let normaliseWhite = pretty && !Element.preserveWhitespace(parent);
-    let stripWhite = pretty && NodeUtils.isDocument(parent);
+    let stripWhite = pretty && Document.is(parent);
     Entities.escapeImpl( accum,  this.coreVal,  setting, false,  normaliseWhite, stripWhite   );
   }
 
